app/requirements.txt
Flask==3.0.0
redis==5.0.1
python-dotenv==0.21.1
